added dto to login has permissions

+++ src/use-cases/logins-has-permissions/delete/remove-permission-delete.usecase.ts
@@ -28,11 +28,21 @@ class RemovePermissionDeleteUseCase {
       throw new RecordNotFound("permission");
     }
 
-    return await this.loginHasPermissionRepository.removePermission({
+    const permissions = await this.loginHasPermissionRepository.loginHasPermission({
+      uuid_login:userExists.uuid,
+      uuid_permission:permissionExists.uuid
+    })
+
+    if(!permissions) {
+      throw new RecordNotFound("permissions");
+    }
+
+    await this.loginHasPermissionRepository.removePermission({
       uuid_login: userExists.uuid,
       uuid_permission: permissionExists.uuid
     })
 
+    return permissions;
   }
 }
